+++ server/clientRouter.jsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { renderToString } from 'react-dom/server'
 import { Provider } from 'react-redux'
-import { routes, configureStore } from '../shared'
+import routes from '../src/routes'
+import configureStore from '../src/configureStore'
 import { RouterContext, match } from 'react-router'
 import manifest from '../dist/ui/manifest.json'
 import qs from 'qs'
