--- shared/index.js
@@ -1,2 +1,0 @@
-export { default as routes } from './routes'
-export { default as configureStore } from './configureStore'
