+++ src/components/Main/Main.jsx
@@ -1,17 +1,12 @@
 import React from 'react'
 import { Link } from 'react-router'
 
-import '../../css/application.css'
+import '../../application.css'
 
 class Main extends React.Component {
-  constructor(props) {
-    super(props)
-  }
-
   render() {
     return (
       <div>
-        {/* Add navbar */}
         <Link to="/home">Home</Link>
         {this.props.children}
       </div>
