+++ src/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
-import { routes } from '../shared'
-import { configureStore } from '../shared'
+import routes from './routes'
+import configureStore from './configureStore'
 
 const initialState = window.__INITIAL_STATE__,
       store = configureStore(initialState),
